base working version
Stable version with a given package architecture. Jdbc implementation for product, swagger w/o custom OpenAPI, base migrations and validation are ready
