+++ services/product-microservice/db/src/main/resources/db/query/product/delete_by_id.sql
@@ -1,0 +1,1 @@
+DELETE FROM product WHERE id = ?
