+++ services/product-microservice/db/src/main/resources/db/query/product/find_by_id.sql
@@ -1,0 +1,1 @@
+select * from product where id = ?
