+++ services/product-microservice/db/src/main/resources/db/query/product/save.sql
@@ -1,0 +1,3 @@
+INSERT INTO product (title, description, photo_url, price, category, quantity)
+VALUES (?, ?, ?, ?, ?, ?)
+RETURNING *;
