+++ services/product-microservice/db/src/main/resources/db/query/product/update_quantity.sql
@@ -1,0 +1,5 @@
+UPDATE product
+SET quantity = quantity + v.delta
+    FROM (VALUES (?, ?)) AS v(product_id, delta)
+WHERE product.id = v.product_id
+  AND product.quantity + v.delta >= 0;
