+++ services/product-microservice/impl/src/main/java/com/technokratos/eateasy/product/ProductApplication.java
@@ -1,0 +1,13 @@
+package com.technokratos.eateasy.product;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = "com.technokratos.eateasy.product")
+public class ProductApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProductApplication.class, args);
+	}
+
+}
