+++ services/product-microservice/impl/src/main/java/com/technokratos/eateasy/product/ServletInitializer.java
@@ -1,0 +1,13 @@
+package com.technokratos.eateasy.product;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(ProductApplication.class);
+	}
+
+}
