+++ services/product-microservice/settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'product-microservice'
+include 'api'
+include 'db'
+include 'impl'
+
