--- services/product-microservice/src/main/java/com/smesitejl/product/ProductApplication.java
@@ -1,13 +1,0 @@
-package com.smesitejl.product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProductApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProductApplication.class, args);
-	}
-
-}
