--- services/product-microservice/src/main/java/com/smesitejl/product/ServletInitializer.java
@@ -1,13 +1,0 @@
-package com.smesitejl.product;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(ProductApplication.class);
-	}
-
-}
