--- services/product-microservice/src/main/java/com/smesitejl/product/cfg/SwaggerConfig.java
@@ -1,17 +1,0 @@
-package com.smesitejl.product.cfg;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info().title("Product API")
-                        .description("API for product microservice")
-                        .version("v1"));
-    }
-}
