--- services/product-microservice/src/main/java/com/smesitejl/product/controller/ProductController.java
@@ -1,44 +1,0 @@
-package com.smesitejl.product.controller;
-
-import com.smesitejl.product.api.ProductApi;
-import com.smesitejl.product.entity.Product;
-import com.smesitejl.product.service.ProductService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.UUID;
-
-@RestController
-@RequiredArgsConstructor
-public class ProductController implements ProductApi {
-
-    private final ProductService productService;
-
-    @Override
-    public ResponseEntity<Product> getById(UUID productId) {
-        return productService.getById(productId);
-    }
-
-    @Override
-    public ResponseEntity<Product> create(@Valid @RequestBody Product product) {
-        return productService.create(product);
-    }
-
-    @Override
-    public ResponseEntity<Void> updateQuantity(UUID productId, @Valid @RequestBody Integer quantity) {
-        return productService.updateQuantity(productId, quantity);
-    }
-
-    @Override
-    public ResponseEntity<Void> update(UUID productId, @Valid @RequestBody Product product) {
-        return productService.update(productId, product);
-    }
-
-    @Override
-    public ResponseEntity<Void> delete(UUID productId) {
-        return productService.delete(productId);
-    }
-}
