--- services/product-microservice/src/main/java/com/smesitejl/product/exception/GlobalExceptionHandler.java
@@ -1,64 +1,0 @@
-package com.smesitejl.product.exception;
-
-import jakarta.validation.ConstraintViolationException;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    // Обработка ошибок валидации, вызванных аннотациями в модели
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, Object>> handleValidationErrors(MethodArgumentNotValidException ex) {
-        Map<String, Object> response = new LinkedHashMap<>();
-        response.put("error", "Validation failed");
-
-        List<String> details = ex.getBindingResult().getFieldErrors().stream()
-                .map(err -> err.getField() + ": " + err.getDefaultMessage())
-                .toList();
-
-        response.put("details", details);
-        return ResponseEntity.badRequest().body(response);
-    }
-
-    // Обработка ошибок валидации, вызванных аннотациями валидации объектов
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity<Map<String, Object>> handleConstraintViolation(ConstraintViolationException ex) {
-        Map<String, Object> response = new LinkedHashMap<>();
-        response.put("error", "Validation failed");
-
-        List<String> details = ex.getConstraintViolations().stream()
-                .map(v -> v.getPropertyPath() + ": " + v.getMessage())
-                .toList();
-
-        response.put("details", details);
-        return ResponseEntity.badRequest().body(response);
-    }
-
-    // Обработка ошибок нарушения ограничений БД
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<Map<String, Object>> handleDataIntegrity(DataIntegrityViolationException ex) {
-        Map<String, Object> response = new HashMap<>();
-        response.put("error", "Conflict");
-        response.put("message", "Database constraint violated");
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(response);
-    }
-
-    // Обработка ошибок, которые не были перехвачены более специфичными обработчиками
-    @ExceptionHandler(Throwable.class)
-    public ResponseEntity<Map<String, Object>> handleGeneralError(Throwable ex) {
-        Map<String, Object> response = new HashMap<>();
-        response.put("error", "Internal Server Error");
-        response.put("message", ex.getMessage());
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
-    }
-}
