--- services/product-microservice/src/main/java/com/smesitejl/product/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package com.smesitejl.product.repository;
-
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository {
-
-}
