--- services/product-microservice/src/main/java/com/smesitejl/product/validation/ValidProduct.java
@@ -1,17 +1,0 @@
-package com.smesitejl.product.validation;
-
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.*;
-
-@Documented
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ProductValidator.class)
-public @interface ValidProduct {
-    String message() default "Product validation failed";
-    Class<? extends Payload>[] payload() default {};
-
-}
