--- services/product-microservice/src/main/resources/application.yaml
@@ -1,41 +1,0 @@
-server:
-  port: 8081
-
-spring:
-  datasource:
-    url: jdbc:postgresql://localhost:5432/product
-    username: eat_easy
-    password: easy
-    driver-class-name: org.postgresql.Driver
-
-  liquibase:
-    enabled: true
-    change-log: classpath:db/changelog/db.changelog-1.0.yaml
-
-  jpa:
-    hibernate:
-      ddl-auto: none
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-        hibernate.hibernate.connection.isolation: 2
-    database-platform: org.hibernate.dialect.PostgreSQLDialect
-
-  jackson:
-    serialization:
-      indent_output: true
-
-  springdoc:
-    api-docs:
-      enabled: true
-    swagger-ui:
-      enabled: true
-      path: /swagger-ui.html
-
-logging:
-  level:
-    root: info
-    org.hibernate.SQL: debug
-    org.springdoc: debug
-    org.hibernate.type.descriptor.sql.BasicBinder: trace
