--- services/product-microservice/src/main/resources/db/changelog/db.changelog-1.0.yaml
@@ -1,51 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: 1
-      author: Smesitejl
-      changes:
-        - createTable:
-            tableName: product
-            columns:
-              - column:
-                  name: id
-                  type: UUID
-                  constraints:
-                    primaryKey: true
-                    nullable: false
-              - column:
-                  name: title
-                  type: varchar(255)
-                  constraints:
-                    nullable: false
-              - column:
-                  name: description
-                  type: varchar(1000)
-              - column:
-                  name: photo_url
-                  type: varchar(255)
-              - column:
-                  name: price
-                  type: decimal(10,2)
-                  constraints:
-                    nullable: false
-              - column:
-                  name: category
-                  type: varchar(255)
-                  constraints:
-                    nullable: false
-              - column:
-                  name: quantity
-                  type: int
-                  constraints:
-                    nullable: false
-
-  - changeSet:
-      id: 2
-      author: Smesitejl
-      changes:
-        - createIndex:
-            indexName: idx_product_title
-            tableName: product
-            columns:
-              - column:
-                  name: title
