--- services/product-microservice/src/test/java/com/smesitejl/product/ProductApplicationTests.java
@@ -1,13 +1,0 @@
-package com.smesitejl.product;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProductApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
